requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
discord.py==1.2.5
idna==2.8
lavalink==3.0.0
multidict==4.7.4
mysql-connector==2.2.9
python-dotenv==0.10.5
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.8
websockets==6.0
yarl==1.4.2
